refactor: rearranging files in the test-commons package
